Divergent
Class Point {

double x;

double y;

double z;

Point(x,y,z);

}

Class Box {

string name;

Point a;

Point b;

Box(Point a, Point b);

}

a. Write a function that takes in a Box and a point, and returns rather or not the point lies within the Box.

b. Write a function that takes in two Boxes and returns whether the Boxes intersect.

c. Write a function that takes in a Box and a point, and returns whether a point is on the faces of the Box.

d. Input: a super large set of boxes and a small set of points, write an optimized function to return an array of point/box(es) association.